把type、state2个字段由Byte修改为Integer类型

+++ miner/miner-server/src/main/java/com/mei/hui/miner/entity/SwarmTicket.java
@@ -14,14 +14,29 @@ import java.util.Date;
 public class SwarmTicket {
     private Long id;
 
+    /**
+     * 节点ip
+     */
     private String nodeIp;
 
+    /**
+     * 节点port
+     */
     private Integer nodePort;
 
-    private Byte type;
+    /**
+     * 类型:0-无效;1-有效
+     */
+    private Integer type;
 
-    private Byte state;
+    /**
+     * 状态:0-未兑换；1-兑换成功；2-兑换失败；3-兑换中
+     */
+    private Integer state;
 
+    /**
+     * 票的面值
+     */
     private BigDecimal parValue;
 
     private String peer;
